Rename: MainPage에서만 사용하는 organisms 컴포넌트들 폴더 분리 및 경로 이동
